Track Time Interval Instead of Timeout
To derive the score, the remaining time on timer has to be tracked. Hence, changing implementation from using setTimeout to control game states, to utilising setInterval to capture time remaining on the respective timer. The game state can then be derived from the remaining time.

Result modal to also show the computed score based on time remaining if the player stop the timer on time. The remaining time state is to be reset on closing the result modal.