Implement Tasks Component for Project
Created projects can have multiple new tasks added, thus, a tasks field is added to the projectsState to store the tasks created for the respective project.

Current implementation is not ideal, as function for task creation and deletion (to be implemented) is being passed over through multiple components resulting in props drilling. This will be fixed subsequently.

*note: exercise used array to store all the tasks for all projects, however, current implementation utilise hashmap/dict (indexed by projectId) to optimise accessing tasks from the respective project on selection and deletion.