Utilise Component Composition to Avoid Prop Drilling
Prop drilling is present due to the shoppingCart state is required (for accessing or manipulating) in nested child components in different locations.

To fix this by embracing component composistion, pull out code that is responsible for rendering elements which requires accessing or manipulating the shoppingCart state on App component and pass them down through the "children" prop. Thus, removing the need for passing down props for nested components, avoiding prop drilling.

Though, it may not be ideal as the app component would become bloated as it requires handling this additional responsibility.