Removed Redundant useEffect Utilisation
useEffect is not required in the use case where the pickedPlaces state is set by retrieving places from the local storage.

This is because the process of retrieving storedPlaces from local storage is synchronous, the execution order is guaranteed. Thus, storedPlaces can be retrieved prior to the pickedPlace state definition, and use to define the state instead.