Fix: Bug on Modal Open State If Esc Button Is Pressed
If esc button is used to close the modal, the modal open state is not set to false. Resulting in the state no longer in sync and modal cannot be open as re-render will not be triggered.

Handle the action of closing the modal with esc button, setting modal open state to false.