Auto Confirm Remove on Timer
Using timer to automatically confirm removal with useEffect hook.

Additionally, returned function in a useEffect hook is used as clean up function, it is executed on component unmount (removed from DOM) or before the next execution of the call back function in the useEffect hook.

Current implementation could result in indefinite loop, as the dependency array consist of a function reference. the pointer to the function reference can change (on redefinition) when the component that defines the function get rerendered.