Fix: Ensure Call Back Function Reference Pointer Consistency
Utilising useCallback to define the dependent call back function (for the useEffect hook), to ensure the function will not be redefined on every rerender on the defining component. This will prevent the potential issue with indefinite loop due to the function reference keep changing on rerender.

Similarily, the dependency array in the useCallback hook should contain the states and props (or any other dependent variables).